+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Enh #564: Clarify `$relations` parameter type in `JoinWith::__construct()` from `array<string|Closure>` to
   `array<string|callable(ActiveQueryInterface):void>` (@vjik)
 - Bug #561: Fix `ActiveRecordInterface::upsert()` with `$updateProperties = false` (@Tigrov)
+- Bug #550: Relation query should be created by related class, not primary model class (@batyrmastyr)
 
 ## 1.0.2 March 11, 2026