+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Yii Active Record Change Log
 
-## 1.0.2 under development
+## 1.0.3 under development
+
+- Bug #558: Fix `SoftDelete` with initiated custom date (@Tigrov)
+- Enh #564: Clarify `$relations` parameter type in `JoinWith::__construct()` from `array<string|Closure>` to
+  `array<string|callable(ActiveQueryInterface):void>` (@vjik)
+- Bug #562: Fix `ActiveRecordInterface::upsert()` to prioritize passed associative values during updates (@Tigrov)
+- Bug #561: Fix `ActiveRecordInterface::upsert()` with `$updateProperties = false` (@Tigrov)
+- Bug #550: Relation query should be created by related class, not primary model class (@batyrmastyr)
+
+## 1.0.2 March 11, 2026
 
 - Bug #544: Revert changes from #538 related to create model without a constructor (@Tigrov)